Added authentication check and authorization token to httpClient

+++ RecipesExercises/RecipeExercise9/RecipeExercise9/Services/RecipeService.cs
@@ -1,4 +1,7 @@
-﻿using System.Net.Http.Json;
+﻿using Microsoft.JSInterop;
+using System.Net.Http.Headers;
+using System.Net.Http.Json;
+
 namespace RecipeExercise9;
 public interface IRecipeService
 {
@@ -11,14 +14,30 @@ public interface IRecipeService
 public class RecipeService : IRecipeService
 {
     private readonly HttpClient _httpClient;
-    public RecipeService(HttpClient httpClient)
+    private readonly IJSRuntime _jsRuntime;
+
+    public RecipeService(HttpClient httpClient, IJSRuntime jsRuntime)
     {
         _httpClient = httpClient;
+        _jsRuntime = jsRuntime;
     }
+
+    public async Task<bool> CheckAuthentication()
+    {
+        var token = await _jsRuntime.InvokeAsync<string>("sessionStorage.getItem", "token");
+        _httpClient.DefaultRequestHeaders.Authorization =
+            new AuthenticationHeaderValue("bearer", token);
+
+        return !string.IsNullOrEmpty(token);
+    }
+
     public async Task<List<Recipe>> GetRecipes()
     {
         try
         {
+            if (!await CheckAuthentication())
+                throw (new Exception("Not authenticated"));
+
             var recipes =  await _httpClient.GetFromJsonAsync<List<Recipe>>("recipes");
             if (recipes is null)
                 return new List<Recipe>();
@@ -29,10 +48,13 @@ public class RecipeService : IRecipeService
             throw new Exception(ex.Message);
         }
     }
+
     public async Task<HttpResponseMessage> AddRecipe(Recipe recipe)
     {
         try
         {
+            if (!await CheckAuthentication())
+                throw (new Exception("Not authenticated"));
             return await _httpClient.PostAsJsonAsync("recipes", recipe);
         }
         catch (Exception ex)
@@ -40,10 +62,13 @@ public class RecipeService : IRecipeService
             throw new Exception(ex.Message);
         }
     }
+
     public async Task<HttpResponseMessage> UpdateRecipe(Recipe recipe)
     {
         try
         {
+            if (!await CheckAuthentication())
+                throw (new Exception("Not authenticated"));
             return await _httpClient.PutAsJsonAsync($"recipes", recipe);
         }
         catch (Exception ex)
@@ -51,10 +76,13 @@ public class RecipeService : IRecipeService
             throw new Exception(ex.Message);
         }
     }
+
     public async Task<HttpResponseMessage> DeleteRecipe(int id)
     {
         try
         {
+            if (!await CheckAuthentication())
+                throw (new Exception("Not authenticated"));
             return await _httpClient.DeleteAsync($"recipes/{id}");
         }
         catch (Exception ex)
